+++ CHANGELOG.md
@@ -33,6 +33,7 @@ doivent être mises à jour #1667
 - La classe `TdtTypologieChangeByApi` est déplacée Pastell\Step\Tdt\Acte\Action\TypologieOnChangeAction #2283
 - Il n'est plus possible d'utiliser l'adresse mail dans la page "Mot de passe oublié ?" #2309
 - Les droits 'changement-email:' ne sont plus sur le rôle admin par défaut #2320
+- Renommage de ChorusProImportUtilService::getMinDateStatutCourant en getOldestDateDepuisLe #2340
 
 ## Suppressions